requirements.txt
flask==1.1.2
Werkzeug==0.16.1
flask_restplus==0.13.0
Flask_Cors==3.0.10
torch==1.7.1
torchvision==0.8.2
Pillow==8.2.0
py-eureka-client==0.9.8
wandb==0.10.33
scikit-image==0.18.1
pytorch-lightning==1.3.2
matplotlib==3.4.1
